Change name space
`Policies::DataRetention` overloads `Policies` which is already heavily
used in the app. This commit removes the `Policies` namespace from data
retention, promotes `DataRetention` to it's place and renames
`DataRetention::Policy` to `DataRetention::RetentionSchedule` which more
accruatley names what these classes do.
