+++ app/models/data_retention.rb
@@ -1,4 +1,4 @@
-module Policies::DataRetention
+module DataRetention
   def self.table_name_prefix
     "policies_data_retention_"
   end
