+++ app/models/data_retention/recovery.rb
@@ -1,0 +1,5 @@
+module DataRetention
+  class Recovery < ApplicationRecord
+    belongs_to :claim
+  end
+end
