--- app/models/policies/data_retention/recovery.rb
@@ -1,7 +1,0 @@
-module Policies
-  module DataRetention
-    class Recovery < ApplicationRecord
-      belongs_to :claim
-    end
-  end
-end
