+++ app/models/policies/further_education_payments/data_retention/retention_schedule.rb
@@ -1,7 +1,7 @@
 module Policies
-  module InternationalRelocationPayments
+  module FurtherEducationPayments
     module DataRetention
-      class Policy < Policies::DataRetention::BasePolicy
+      class RetentionSchedule < ::DataRetention::BaseSchedule
         self.claim_attributes = {
           id: :retained,
           academic_year: :retained,
@@ -51,7 +51,7 @@ module Policies
           onelogin_idv_full_name: :inactive_claim_submitted_in_prior_academic_term?,
           onelogin_idv_last_name: :inactive_claim_submitted_in_prior_academic_term?,
           onelogin_idv_return_codes: :inactive_claim_submitted_in_prior_academic_term?,
-          onelogin_uid: :inactive_claim_submitted_in_prior_academic_term?,
+          onelogin_uid: :inactive_claim_over_six_years_old?,
           onelogin_user_info: :inactive_claim_submitted_in_prior_academic_term?,
           paye_reference: :inactive_claim_submitted_in_prior_academic_term?,
           payroll_gender: :retained,
@@ -87,25 +87,77 @@ module Policies
 
         self.eligibility_attributes = {
           id: :retained,
-          created_at: :retained,
-          updated_at: :retained,
-          application_route: :retained,
           award_amount: :retained,
-          breaks_in_employment: :retained,
-          changed_workplace_or_new_contract: :retained,
-          current_school_id: :retained,
-          date_of_entry: :retained,
-          employment_history: :retained,
-          nationality: :retained,
-          one_year: :retained,
-          passport_number: :inactive_claim_over_two_years_old?,
-          previous_year_claim_ids: :retained,
+          building_construction_courses: :retained,
+          chemistry_courses: :retained,
+          claimant_date_of_birth: :inactive_claim_submitted_in_prior_academic_term?,
+          claimant_identity_verified_at: :retained,
+          claimant_national_insurance_number: :inactive_claim_submitted_in_prior_academic_term?,
+          claimant_passport_number: :inactive_claim_submitted_in_prior_academic_term?,
+          claimant_postcode: :inactive_claim_submitted_in_prior_academic_term?,
+          claimant_valid_passport: :retained,
+          computing_courses: :retained,
+          contract_type: :retained,
+          created_at: :retained,
+          early_years_courses: :retained,
+          engineering_manufacturing_courses: :retained,
+          fixed_term_full_year: :retained,
+          flagged_as_duplicate: :retained,
+          flagged_as_mismatch_on_teaching_start_year: :retained,
+          flagged_as_previously_start_year_matches_claim_false: :retained,
+          further_education_teaching_start_year: :retained,
+          half_teaching_hours: :retained,
+          hours_teaching_eligible_subjects: :retained,
+          maths_courses: :retained,
+          passport_number: :inactive_claim_submitted_in_prior_academic_term?,
+          physics_courses: :retained,
+          possible_school_id: :retained,
+          provider_assigned_to_id: :inactive_claim_submitted_in_prior_academic_term?,
+          provider_verification_chase_email_last_sent_at: :retained,
+          provider_verification_claimant_bank_details_match: :retained,
+          provider_verification_claimant_date_of_birth: :inactive_claim_submitted_in_prior_academic_term?,
+          provider_verification_claimant_email: :inactive_claim_submitted_in_prior_academic_term?,
+          provider_verification_claimant_employed_by_college: :retained,
+          provider_verification_claimant_employment_check_declaration: :retained,
+          provider_verification_claimant_national_insurance_number: :inactive_claim_submitted_in_prior_academic_term?,
+          provider_verification_claimant_postcode: :inactive_claim_submitted_in_prior_academic_term?,
+          provider_verification_completed_at: :retained,
+          provider_verification_continued_employment: :retained,
+          provider_verification_contract_covers_full_academic_year: :retained,
+          provider_verification_contract_type: :retained,
+          provider_verification_deadline: :retained,
+          provider_verification_declaration: :retained,
+          provider_verification_disciplinary_action: :retained,
+          provider_verification_email_count: :retained,
+          provider_verification_email_last_sent_at: :retained,
+          provider_verification_half_teaching_hours: :retained,
+          provider_verification_half_timetabled_teaching_time: :retained,
+          provider_verification_not_started_qualification_reason_other: :retained,
+          provider_verification_not_started_qualification_reasons: :retained,
+          provider_verification_performance_measures: :retained,
+          provider_verification_started_at: :retained,
+          provider_verification_taught_at_least_one_academic_term: :retained,
+          provider_verification_teaching_hours_per_week: :retained,
+          provider_verification_teaching_qualification: :retained,
+          provider_verification_teaching_responsibilities: :retained,
+          provider_verification_teaching_start_year: :retained,
+          provider_verification_verified_by_id: :retained,
+          provision_search: :retained,
           redacted_attributes: :retained,
-          school_headteacher_name: :inactive_claim_over_two_years_old?,
-          start_date: :retained,
-          state_funded_secondary_school: :retained,
-          subject: :retained,
-          visa_type: :retained
+          school_id: :retained,
+          subject_to_disciplinary_action: :retained,
+          subject_to_formal_performance_action: :retained,
+          subjects_taught: :retained,
+          taught_at_least_one_term: :retained,
+          teacher_reference_number: :inactive_claim_over_six_years_old?,
+          teaching_hours_per_week: :retained,
+          teaching_qualification: :retained,
+          teaching_responsibilities: :retained,
+          updated_at: :retained,
+          valid_passport: :retained,
+          verification: :inactive_claim_submitted_in_prior_academic_term?,
+          work_email: :inactive_claim_submitted_in_prior_academic_term?,
+          work_email_verified: :retained
         }
       end
     end
