+++ spec/jobs/data_retention/remove_recovery_records_job_spec.rb
@@ -1,12 +1,12 @@
 require "rails_helper"
 
-RSpec.describe Policies::DataRetention::RemoveRecoveryRecordsJob do
+RSpec.describe DataRetention::RemoveRecoveryRecordsJob do
   describe "#perform" do
     let(:claim) { create(:claim) }
 
     context "when the recovery record is due to be destroyed" do
       it "deletes the record" do
-        recovery = Policies::DataRetention::Recovery.create!(
+        recovery = DataRetention::Recovery.create!(
           claim: claim,
           payload: [],
           destroy_at: Date.yesterday.end_of_day
@@ -22,7 +22,7 @@ RSpec.describe Policies::DataRetention::RemoveRecoveryRecordsJob do
 
     context "when the recovery record is not due to be destroyed" do
       it "does not delete the record" do
-        Policies::DataRetention::Recovery.create!(
+        DataRetention::Recovery.create!(
           claim: claim,
           payload: [],
           destroy_at: Date.today.end_of_day
@@ -32,7 +32,7 @@ RSpec.describe Policies::DataRetention::RemoveRecoveryRecordsJob do
           described_class.perform_now
         end
 
-        expect(Policies::DataRetention::Recovery.count).to be(1)
+        expect(DataRetention::Recovery.count).to be(1)
       end
     end
   end
